Add files via upload
This is the "better" version of the script that I'm working on. Technically it's worse at the moment, as it doesn't handle errors properly. When I figure out what's wrong with it, it will be better in that it should provide a more detailed description of failures/successes. 

The final version should check the API key and Hash validity and return errors such as "Invalid API Key" or "404" errors. 

I've removed some of these specifics from this copy of this script while debugging.